Crouch Bug fixed + No jump while crouching + QOL
Changed OnCollider events (from line 49 onwards) into OnTrigger because I set them to Collider incorrectly in the previous commits. This fixed the issue of the player not seeing the ceiling while crouching and clipping into the ground.

Made jumping while crouching impossible - Will have to revise this later to make sure that the controls still feel smooth

Added an extra space just to make the code look a bit cleaner